Pim: Change Content's content field to be string (it is).
Like 22fccf6d9e8ad5d1f51ebb891561bf26ee1b937e this was a poor attempt to future-proof the
database in case we would like to change to encrypted blobs at some point, instead of
saving in the clear.

This was causing issues and is not needed yet, so it was changed.