Example Systemd Unit installs & starts etesync-dav
Use systemd restrictions to isolate the etesync-dav process from the host
system in an attempt to protect the host from bad code in etesync-dav or
its dependencies. Startup is slower as the latest version of etesync-dav
is installed at each run. Startup is even slower after a reboot as the
etesync and code caches are in tmpfs. A Readme provides a FAQ.