Disable the etesync cache to prevent race conditions
The reason for that, is that with PeeWee, the database file is actually
saved per process rather than per EteSync handle due to the database
proxy. Since we were caching the etesync object, we were not reiniting
the database, causing a race condition that would cause data to be saved
on the wrong db.