Support paths (and UIDs) that contain front slashes (/)
Apparently it's a legal characters in UIDs and it can cause radicale to
fail. Unfortunately this is not theoretical. Outlook generated UIDs look
like base64 encoded strings, which in turn means they may contain /, so
this is actually a real issue.

This not really a fix, but a workaround. What we do is rewrite the URL
(and UID) and replace every "/" with a "," which can't appear in a UID
so we know it's safe, and then we do the reverse the other way around.

This solution however is not perfect as Radicale doesn't really support
rewriting URLs, so when e.g. importing such events in Thunderbird you
may need to wait until the next refresh (or trigger one manually) in
order to see your changes.

Fixes https://github.com/etesync/etesync-dav/issues/73