Partial revert "Support paths (and UIDs) that contain front slashes (/)"
This was a great workaround, but I'm going to fix it differently now by
just generating hrefs for each item rather than using the uid.

This is not a complete revert, as we're keeping the part that limits the
depth of an item and rewrite the path if too deep. Needed when importing
an item with a / in the uid using Thunderbird.

Ref: https://github.com/etesync/etesync-dav/issues/73

This reverts commit 4e8f61a7b2542ce32e2cc7192e97e04c0acc949a.