Filter out the PHOTO field from vCards until we fix handling.
Radicale and vObject (the underlying library) don't support vCard 4.0
which makes vCards with PHOTO serialize wrongly and break the sync. This
commit filters out this field from vCards so at least sync is fixed
until vCard 4.0 support is implemented.

Important: editing a contact through etesync-dav will cause that contact
to lose its photo if it had one. The problem is almost identical to the
one this commit works around: etesync uses the vCard format internally
and the underlying library (vObject) doesn't serialize this field
correctly.

This is a workaround for #65.