Transform vCard 4.0 contacts to 3.0 when serving
This is a workaround for both #70 and #66. Essentially the problem is
that both Radicale and macOS only support vcard 3.0 and EteSync uses
version 4.0 internally.
Luckily, version 3.0 and 4.0 are quite similar, so we can hackily
transform from 4.0 to 3.0 when serving contacts so they at least show
and work on macOS.

We can in the future actually implement correct transformations, though
at least for now, this works and solves the issues people were facing on
Macs.