Sync with etesync asynchronously instead of blocking requests.
This was causing some clients to timeout and was anyway causing for
bad performance. With this patch the sync happens in the background
and (almost) doesn't block requests so it behaves much better.

The main problem with this patch is that while it pushes changes
immediately, users will only see pulled changes after the next sync.

If for example you have two clients, the Android one and the DAV one
which are fully in sync.
Now assume Android made some changes, and now you want to make a
change to an event in Thunderbird. Thunderbird will now push these
changes to DAV which will sync with the server **immediately after**
the Thunderbird request. So the EteSync server will get the new
event, but Thunderbird will only get to see changes after the next
time it syncs. So essentially Thunderbird would need to sync twice
to see new changes from the server.

This is an attempt at fixing #93.