Add NamedReverseSemaphore: a reverse semaphore that's key based
If we are holding a lock on a key already, the same key can be used
again to access the resource without waiting. If a different key is
encountered, it waits until it can get the lock itself.
It lets us just access the same etesync object without waiting if we
are already using it, but we lock if we want to get the etesync object
of another user.