Storage: fix locking around the storage and etesync cache
We weren't correctly locking around the sync thread, causing mulitple
sync threads to be started at the same time. Additionally, because we
were using the NamedReverseSemaphore when getting the etesync object, we
could have a race and get it twice, which would cause two sync threads
to be ran and thus could cause a lock and a conflict.

Fixes #132