Sync: change the sync back to being synchronous instead of async.
We made it async in the first place in order to be able to sync
automatically for users so they don't have to wait when the sync
operation is slow. It actually makes more sense to just have a short
timeout (in our case, 5 seconds) and continue the request with stale
data if it takes too long.
It's the best of both worlds, though the only problem that comes to mind
is what happens if there are consecutive requests which all will wait on
the timeout, though I guess that's up to the dav client to handle.