Don't force binding to ipv6 when binding to localhost.
It was originally added because Radicale didn't know how to deal with ipv6
automatically, though it's no longer the case with v3. Radicale now
bindings to both ipv4 and 6 automatically if localhost supports it.

We are reverting it because it was causing issues for users who don't
have ipv6. Potentially also fixes #151

Partial revert of 944f2fa2f1674ec657937f7a0bf6ef563579393e