macOS: run a mainloop so that activity monitory knows we are running.
Activity Monitory on macOS requires apps to report to it that they are
still running. So we now run an event loop in a thread in order to fix
this issue.

There's actually an issue with PyObjC that makes this change not work
unless you use PyObjC from git. Commit with the fix:
https://github.com/ronaldoussoren/pyobjc/commit/34d59f31e37a318b2374f019246bf10bdbb28acb

This fixes #87 and possible also #142.