Fix race condition when there are many connections in parallel.
This was happening quite consistently for me with Evolution with an
account with many collections.

The problem was that the locking mechanism we had was locking to
make sure that the peewee database isn't being changed, but it
wasn't ensuring that the self.etesync of the storage object isn't
being reset. We added a lock and a refcount to fix it.

Fixes #173.