Sync: fix syncing of item deletions.
They weren't being synced because we were returning sync tokens as if we
support them, but we actually don't.
When sync tokens are enabled all of the items that changed (including
deleted) should be returned, while in our case we were omitting deleted
ones because those aren't returned when not using sync tokens.

The solution is thus to disable sync tokens support again.